% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob-package.R
\docType{package}
\name{blob-package}
\alias{blob-package}
\title{blob: A Simple S3 Class for Representing Vectors of Binary Data ('BLOBS')}
\description{
R's raw vector is useful for storing a single binary object. What if you want to put a vector of them in a data frame? The 'blob' package provides the blob object, a list of raw vectors, suitable for use as a column in data frame.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://blob.tidyverse.org}
  \item \url{https://github.com/tidyverse/blob}
  \item Report bugs at \url{https://github.com/tidyverse/blob/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com}

Authors:
\itemize{
  \item Hadley Wickham
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
