% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{blob}
\alias{blob}
\alias{new_blob}
\alias{validate_blob}
\alias{as_blob}
\alias{is_blob}
\title{Construct a blob object}
\usage{
blob(...)

new_blob(x = list())

validate_blob(x)

as_blob(x)

is_blob(x)
}
\arguments{
\item{...}{Individual raw vectors}

\item{x}{A list of raw vectors, or other object to coerce}
}
\description{
\code{new_blob()} is a low-level constructor that takes a list of raw vectors.
\code{blob()} constructs a blob from individual raw vectors.
\code{as_blob()} and \code{is_blob()} are simple forwarders to \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[=inherits]{inherits()}}, respectively.
}
\examples{
x1 <- charToRaw("Good morning")
x2 <- as.raw(c(0x48, 0x65, 0x6c, 0x6c, 0x6f))

new_blob(list(x1, x2))
blob(x1, x2)

as.blob(c("Good morning", "Good evening"))
}
\seealso{
\code{\link[=as.blob]{as.blob()}} for the legacy interface for specifying casts.
}
