% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{vec_type2.vctrs_blob}
\alias{vec_type2.vctrs_blob}
\title{Coercion}
\usage{
\method{vec_type2}{vctrs_blob}(x, y)
}
\arguments{
\item{x}{Either vector types; i.e.}

\item{y}{Either vector types; i.e.}
}
\description{
Double dispatch methods to support \code{\link[vctrs:vec_type2]{vctrs::vec_type2()}}.
}
