% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{as.blob}
\alias{as.blob}
\title{Deprecated generic}
\usage{
as.blob(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
The \code{as.blob()} generic has been deprecated in favor of
the \code{\link[=as_blob]{as_blob()}} function (for users) and the \code{\link[=vec_cast.blob]{vec_cast.blob()}} method
(for implementers).
Implement a \code{vec_cast.blob.myclass()} method to support
coercing objects of your class to blobs.
See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} for more detail.
}
\keyword{internal}
